CA-325582: Remove SR DSs from memory on SR.forget and destroy
Make sure data sources for an SR are not kept in memory in xcp-rrdd
when the SR is destroyed or forgotten.

sr-destroy is separated into 3 stages: the first is
executed by the SR master, which physically destroys the contents of the
SR, the second is executed by all the hosts that have had a PBD,
which cleans the metrics about them. And the third one is the database
cleanup, executed by the pool master.

sr-forget has been separated into 2 steps, which correspond to the last
2 of sr-destroy

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>