CA-328130 Make sure speed updated after update
Without this patch, USBs that remain plugged in
over the lifetime of an update will have an empty
speed value. An ad-hoc solution for this is to
unplug the USB, `pusb-scan`, plug the USB back in
and `pusb-scan` once more.

Speed was added to the PUSB datamodel, so by
default takes an empty string value. We need to
update this value directly because `pusb-scan`
does not refresh USBs that it already knows
about.

Signed-off-by: lippirk <ben.anson@citrix.com>