CA-325582 fix some storage quicktests for pools
This does not fix all failures, but some.

Some storage quicktests were failing on pools
because by default it runs each test for each SR
a host can see - including the local storage of
another host in the pool. To fix we add a filter
to remove such SRs.

`vdi_snapshot_in_pool` was also iterating over
hosts - it should not have been, because we
don't necessarily have access to that host's SRs.
Instead, only test snapshot on localhost.

Signed-off-by: lippirk <ben.anson@citrix.com>