CP-32863: fix hotplug script for storage driver domains
It was trying to access the hotplug status of the wrong domain (dom0),
instead of the backend domain when storage driver domains are used.

Pass a new environment variable XENBUS_PREFIX that is by default empty, and contains
`/local/domain/<backend-domid/` otherwise.

For the device to get correctly plugged we need to have the
physical-device and physical-device-path xenstore keys written.
Previously these were written based on the contents of xenstore keys `params` (SMAPIv1),
or `qemu-params` (SMAPIv3 via qemuback.py).
When the backend is in a different domain we cannot `stat` the device
though, so we need to get the `major:minor` numbers from the SMAPI
plugin.

Read `qemu-params`, and if it contains an entry in the correct format
(`device-path major:minor`), then write out the other 2 xenstore keys
as part of the same transaction.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>