CP-32863: do not get stuck waiting for state to reach closed when key disappears
With storage driver domains I noticed the xenstore key disappearing, and
xenopsd hanging until a timeout was reached waiting for the state key to
reach state 6 (closed).

If the frontend is completely gone away with all its xenstore keys, then
there is nothing to wait for, and consider the unplug successful.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>