CP-32863: Do not wait for hotplug scripts in Dom0 when plugging in storage driver domains
Storage driver domains' disks aren't accessible to Dom0.
There is logic already in xenopsd to detect that, and plug a PV disk
into Dom0 using the storage driver domain as backend when it needs
access to a block device (instead of accessing the block device directly,
or using Nbd).

The block device cannot be accessed directly, because it doesn't exist
in Dom0 (or even if something happened to exist with same path it would
        be something completely different!).
NBD doesn't currently work due to sandbox restrictions in qemu (it
cannot use TCP networking, just unix domain sockets).
It is possible to get NBD working if we set up a socat proxy
(the SMAPI plugin should do so if it needs to), but the PV ring should
be more efficient, albeit we loose sparseness information that is useful
for VM export.

The hotplug scripts were getting stuck when plugging devices into Dom0
, this isn't what they were designed for: in this case Dom0 acts as a
frontend not as a backend, and we plug in an `/dev/xv*` device, not a
tapdev device.

Skip running of hotplug scripts when the target is Dom0 and the source
is backend driver domain outside of Dom0.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>