CP-32863: fix storage driver domain shutdown
When we detected that a storage driver domain VM was shut down we sent
an Client.SR.reset API call, which determined that this SR/PBD has a
storage driver domain attached that is halted, and triggered a boot.
Effectively ensuring you can never shut it down cleanly.
We now have better protection against this by using allowed operations
(see previous commit), so fix this bug.

Move the PBD plug code into the branch of `match` that runs when the VM
is *not* shutdown.
When the VM is shut down then this branch triggers instead:
`| [] -> () (* happens when the VM is shutdown *)`, and no PBD plug is
attempted.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>