CA-144246 prevent log spam when intellicache enabled
To find the sr+vdi pairs that are in use, the iostat daemon was
searching "/dev/sm/phy" for files which look like
"/dev/sm/phy/$sruuid/$vdiuuid" (during testing it seems that those are
actually symbolic links to "/var/run/sr-mount"). It adds these paths,
along with sr/vdi uuids to a map.

Files representing intellicache vdis don't exist in "/dev/sm/phy", so
these were not added to the map. However, `tap-ctl list` _does_ list
these (as "*/$sruuid/$vdiuuid.vhdcache"). The plugin cross references
the map and the output from tap-ctl, and was complaining that the
".vhdcache" files don't exist - causing the log spam.

The fix is to add the .vhdcache files to the map.

Signed-off-by: lippirk <ben.anson@citrix.com>