CA-334909: Only update if the update is needed
When we take a snapshot of a guest, we store the fields of that guest in
the snap metadata. When we revert to this snapshot, we copy the fields
across, but with the domain type we would call
`ensure_domain_type_is_consistent`. This function works to ensure that
there is always a domain type set in the field, and in the case of an
hvm guest, would also fill it in correctly. However, there are cases
that `derive_domain_type` currently doesn't handle given the HVM boot
policy field. Luckily, when we create the snapshot, we actually store
the domain-type field, so there are cases where we don't need to
_update_ this field, just use it.
The cases where we _do_ need to update this field are if the field is
either missing, or set to `unspecified (snapshots taken on Jura). In
each of these cases, we work out what the domain-type is likely to need
to be, and use that, else we just use what we stored previously

Signed-off-by: Richard Davies <richard.davies@citrix.com>