CA-334951: Use a variant to model recommendations
Before an array could be used to represent possible, impossible
recommendations; as well as only error messages if the list of string
had 3, 4 and 2 elements, respectively.

This changes the xen-api behaviour when an error other than
vm_incompatible happened on a retrieve_wlb_recommendations call.
Before only the 2 strings of the error were returned, now it also
returns the source (WLB), the score (0.0), the id of the consultations
and the 2 strings concatenated into one.
This is the format that vm_incompatible error produced before.

Backport of fac082e73b7576027e69b2f1524fcc52a0fc2a84

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>