CA-333610 modify server.ml to accept InternalAsync calls
InternalAsync calls are designed to behave externally the same as a Sync
call would - the main property we want to preserve is that progress make
during the call is attributed to the task of the original request. For
this reason, we do _not_ create a sub task, and instead execute with the
context handed to us from the master (in the HTTP request). We notify
the caller that we have finished processing their request by completing
the task, hence we must set `need_complete:true`.

The advantage in using an InternalAsync call over a Sync call is that we
do not leave idle connections open.

It should be noted that accepting calls of the form "X_dothing" never
worked with Async, and continutes to not work.

Signed-off-by: lippirk <ben.anson@citrix.com>