CA-334763: Show errors during metadata export/import
Previously the error would've been lost by the thread.
This can occur if for example a control domain is not allowed to be
exported due to a bug, which raises an exception.

The actual bug was fixed by dropping the commit that restricted allowed
operations on system domains: this was not necessary, all driver domains
are control domains now.
In fact the documentation for control domain says: domain 0 or a driver
domain, so there is no need to use system domains for that.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>