CP-33057 rewrite init.d-xapissl script in ocaml
An stunnel systemd service has been written, so use that instead of the
xapissl script to manage stunnel within xapi. At the same time, rewrite
it in ocaml, as it is not necessary for it to be written in bash.

The script can be simplified massively because systemd now handles
problems related to multithreading when starting the daemon. We only
worry about making sure that writing the config file and ssl certificates
to disk is synchronised.

Signed-off-by: lippirk <ben.anson@citrix.com>