Fix cluster_stack_in_use error message
In CA-336253 xe cluster-pool-destroy failed with a confusing error message saying that the cluster stack is already in use.
The reason for the failure was that a PBD unplug has previously failed (but XenRT didn't notice and ignored it), which then caused this failure in destroy.

Improve the error message to say that something earlier failed.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>