CP-32398 must not use POD when using SRIOV vGPU
Add a check into VM start to check its memory configuration if the VM
has an SRIOV vGPU. We disable ballooning (DMC) and POD (populate on
demand) by pinning memory to static max. The main point is disabling POD
but we found it difficult to implement while still permitting
ballooning. So we are disabling both by demanding that the memory
configuration is fixed at the upper limit at the start of the VM.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>