CA-336258: http-svr: add JSONRPC_protocol
I've put this into xmlrpc_client.ml(i), alongside `XML_protocol` and
`XMLRPC_protocol`. The filename does not really make sense anymore, but
changing it would cause a lot of updates (it is used all over the place). And
`XML_protocol` us not XMLRPC either... Not worth doing for a "legacy" module.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>