CA-337113: Turn create_tools_sr into remove_tools_sr
The Tools ISO has been removed from dom0, so there is no need anymore for the
Tools SR.

The removal of the Tools SR is best-effort: a warning is logged if it fails (as
done before for redundant SRs). Failure occurs if there is still some VM that
has a Tools VDI attached to it. The cleanup code runs each time the master xapi
start, so once this situation has been resolved, the SR will eventually get
cleaned up.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>