CA-335964: store the original vm uuid when starting migration
This allows other daemons to hide the temporary uuid from their clients
by reading the value from xenstore.

Since rename is both used on the origin and destination of the
migration and both need different operations on xenstore, a parameter
was added to differentiate between the stages.

The new origin key is written directly to the temporary vm uuid tree,
as writing to the original tree would mean the value gets changed by
the function that moves the vm uuid subtree.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>