CP-33465: xapi-cli-server: initial library
Setup a internal library in ocaml/xapi-cli-server that will become the entire
server-side half of the (xe) CLI. The goal is to restrict the CLI server, such
that it cannot access xapi's internals – it needs to communicate with the main
xapi server entirely using the XenAPI.

Here is the complete list of file that need to be in this library:

- cli_cmdtable.ml
- cli_frontend.ml
- cli_key.ml
- cli_operations.ml
- cli_printer.ml
- cli_progress_bar.ml
- cli_util.ml
- record_util.ml
- records.ml
- xapi_cli.ml

This commit creates the library and dune file, and moves as many files as
possible into it, while not breaking the build (not a lot of files at the
moment). More files will be moved incrementally, after disentangling them from
xapi.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>