CP-32980: set ssid for storage driver domains
This sets the security id for the domain, in a way that will be
compatible with XSM/Flask.

The SSID used here is hardcoded for now, in the future this can be
replaced by looking up and applying a Flask policy when Flask is ready.
A (storage) driver domain is more privileged than a regular DomU, but
less privileged than Dom0.

is_system_domain is true for both storage driver domains and Dom0,
however XAPI never creates Dom0, so the SSID won't apply to it
(it doesn't have to: Dom0 is already a privileged domain).

Signed-off-by: Edwin Török <edvin.torok@citrix.com>