CA-337546: update to new Stunnel_cache API with Safe_resources support
This required moving the stunnel_processes cleanup function to where the
tunnel was opened (convert from [connect] to [with_connect] API).

Purging needs a slight change since we no longer raise Not_found,
but return None instead.

Retrying to open the connection over TCP also needs to be made only
if the connection itself fails, not if the function [f] fails,
hence the use of a fallback variable.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>