CA-336258: Use `is_json:true` in `jsoncallback`
The function `callback1` in `api_server.ml` has a parameter called `is_json`,
and is oddly called with `false` from `jsoncallback`.

We have three HTTP endpoints for RPC requests (see xapi.ml):

    ("post_root", (Http_svr.BufIO (Api_server.callback false)));
    ("post_json", (Http_svr.BufIO (Api_server.callback true)));
    ("post_jsonrpc", (Http_svr.BufIO (Api_server.jsoncallback)));

These map to paths as follows:

    post_root = /
    post_json = /json
    post_jsonrpc = /jsonrpc

The first two have existed "for ever". It seems that /json accepts XML-RPC and
returns JSON in the response, while /jsonrpc is proper JSON-RPC.

The `is_json` parameter of `callback1` may have had slightly different
semantics in the past. The function `jsoncallback` does its own JSONification,
so it probably doesn't rely on `callback1` with `true`. However, now it does,
due to the fix in the previous patch (and on master it already did due to other
changes). Therefore, we need to change it now.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>