CP-33793 make QMP monitoring more robust
The QMP monitoring thread first connects to all existing QMP sockets
that in can find in /var/run/xen. Exceptions here could kill the thread.
This commit makes sure exceptions are caught and logged.

There is still the problem that connecting to a socket could block and
therefore block the entire thread. This could be avoided by implementing
a timeout for the connect() call.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>