CA-334811 assign xapi version automatically
build_info.ml is created during compilation. We tweak this to include
the xapi version. By default this is obtained by a call to git describe,
but can be overriden using a make parameter - this is useful when
creating a xapi rpm.

In order for opam builds to be successful, we must also add a dummy
version environment variable to the opam files.

Signed-off-by: lippirk <ben.anson@citrix.com>