CP-33774 PSR orchestration
We implement a function which orchestrates a pool secret rotation.
We handle failure by splitting the PSR into idempotent chunks, and
saving checkpoints at the end of each chunk. This will allow us to
resume from failed PSRs, and importantly reduce the chance of hosts
becoming locked out with the wrong pool secret.

Signed-off-by: lippirk <ben.anson@citrix.com>