CA-339329 firstboot scripts shouldn't sync DB when ugprading
There are 4 problematic scripts - we fix 2 of them here.

To preserve old behaviour of the firstboot scripts, we don't sync the
DB during an upgrade.

If we do sync the DB during an upgrade, then we get an error complaining
that incoming DBs have an incompatible schema version. This happens
because hosts are told to send the the minimally compliant Miami DB
- see pool_db_backup.ml.

Signed-off-by: lippirk <ben.anson@citrix.com>