Increase sharing of strings in database
Before a string as part of a key or value is added to the database, pass
it trough function share(). It tries to locate the string in the
StringPool to maximise sharing of strings; if not found, a new string is
added to the StringPool. This increases the sharing of strings in the
database: each string should be represented only once and all instances
are using it. Such a string can still be garbage collected when no
instance is reachable any more because the string pool is a weak set.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>