CA-337546: add a xapi_stdext_resources module for safe handling of file descriptors
Prevent double-close of file descriptors and channel pairs.
Automatically close file descriptors based on scope, like a built-in
finally (see Safe.within).

File descriptors can be moved out of a scope for another data structure
to assume ownership of it (e.g. the Stunnel cache), however this
reintroduces the possibility of leaks if used incorrectly.
Track Unixfd.t file descriptors with GC finalisers and close any that we have missed.

It is important to store and plumb through Unixfd.t as far as possible,
and convert to Unix.file_descr only at the last moment before calling
a low-level module so that the leak/double-close detection has a
chance to work.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>