CA-340484 bump xapi version in DB after upgrade/update
We got round to bumping the hardcoded xapi version in b81af6e, but
the xapi version is also stored in the DB, which was not changed.
This was causing problems with the firstboot create-guest-template
service, because ultimately it gets the xapi version from the DB,
and then uses that version to perform a template import. This led
to an IMPORT_INCOMPATIBLE_VERSION error.

Signed-off-by: lippirk <ben.anson@citrix.com>