CA-340776: disconnect from stunnel cleanly
During VM import we are asked to upload a file to a remote URL.
We do that using stunnel by default. However when we closed the
file descriptor stunnel may still have work to do in flushing the data
(needs to wrap it with TLS), so we need to wait for it to exit cleanly
before proceeding.

Otherwise we end up killing stunnel before it has sent the entire
stream, causing the remote end to complain about End_of_file during a VM
import initiated anywhere else than the master.

Previously we just closed the file descriptor and didn't wait for
stunnel, and only killed it when we had an exception.
That still worked, but has the potential of loosing diagnostic
information about stunnel errors, it is better to wait.

Adapted cherry-pick from f15ea33417589436b192fcef49a669f19d5fdcca

Signed-off-by: Edwin Török <edvin.torok@citrix.com>
Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>