CA-341988 don't take basename of empty update key
An update is signed iff its key is non-empty,
otherwise it is unsigned. Rolled-up hotfixes are
considered unsigned.

When parsing update xml files we take the file
name of the key rather than the path, but since
basename "" == ".", xapi erroneously concludes
that rolled-up hotfixes are signed. This causes
problems if you install hotifx A on a host, where
A rolls up B, and then try to install hotifx B on
a different host.

Additionally map any existing update keys in the
DB from "." to "".

Signed-off-by: lippirk <ben.anson@citrix.com>