CP-34379 don't proceed with rotation if PSR state is inconsistent
We add some sanity checks before performing PSR
operations.

The main motivation for this is that if the master
changes after a failed PSR then we want to block
the next PSR. The idea is to avoid lockout at all
costs, even though manual intervention may be
required to fix the issue.

Signed-off-by: lippirk <ben.anson@citrix.com>