Replace Core.Epoll with Polly for epoll bindings
We are using epoll() for QMP communications. The existing implementation
as part of the JaneStreet Core libraries brings along a massive code
base. This commit removes that dependency and brings in a minimal
binding for epoll called Polly. For now, until is proven to be good, it
is vendored under vendor/polly but it could be moved to xs-opam longer
term.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>