CA-341921: avoid EINVAL errors for >1024 FDs
xenopsd supports >1024 FDs, because it uses epoll internally.
However when it calls forkhelpers it can fail, if the forkhelpers FD is
>1024.

We do not actually need to use select here at all, because we are
watching just 1 FD: we can use socket timeouts instead.
(We could also use epoll, but for 1 FD it is probably not worth it)

Signed-off-by: Edwin Török <edvin.torok@citrix.com>