Improve HA parameter derived from timeout
This changes how essential parameters for the HA daemon are derived from
the xe pool-ha-enable timeout parameter. So far, the network heartbeat
was always configured to use 10 heartbeats in an interval. There is good
reason to believe that this contributed to the unreliability of HA.  HA
uses two heartbeats:

(1) UDP network heartbeat - fast and unreliable
(2) State file heartbeat - slow and reliable

HA's network heartbeat is sensitive to network congestion and with only
10 heartbeats configured before a timeout would trigger HA, it was easy
for HA to trigger. This patch addresses the asymmetry of the heartbeats:

* We want more network heartbeats than storage heartbeats because they
  are cheap but can be lost.
* We want a minimum time to pass between heartbeats - less for network
  and more for storage heartbeats.
* We configure many network and 20 storage heartbeats per interval.

To avoid problems with rounding of small integers we do all calculations
in floating point.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>