UPD-666 CA-343117: host-backup: Include /boot/efi in the tarball
When using a UEFI-booted host, /boot/efi is a separate mountpoint and so
it does not get included in the backup tarball due to the use of
--one-file-system. This causes the restore to fail due to missing files.
Explicitly include /boot/efi to fix restore with UEFI hosts. Note that
with legacy-booted hosts, /boot/efi is not a separate mountpoint but
including it twice is fine.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>