maintenance: ensure all fistpoints work as expected
The existence of two fistpoints was only ever checked
when the Xapi_fist module is loaded at runtime, and
that result cached. I.e. from xapi's point of view,
these two fistpoints would be either always present,
or always not present. Instead we want to query the
filesystem each time we check for a fistpoint.

Signed-off-by: lippirk <ben.anson@citrix.com>