CA-322708 - VM must not be allowed to start during storage migration
During VM storage migration it was possible to start the VM.
This leads to errors (VDI in use and illegal storage transition).
To resolve this issue the start and start_on operation were removed from the allowed_operations list in the is_allowed_concurrently function.
This removes the Start and Start On operations from the XenCenter vm menu, thus preventing the user from encountering the errors.

Signed-off-by: Gheorghe Burac <gheorghe.burac@citrix.com>