CP-34357: add designate_new_master to pool operations
The motivation for this is so that we have the information
to avoid running a PSR during a designate new master, and
vice versa. A (good) side effect of this is that designate
new master is blocked during the other pool operations.

To avoid having to assert that a PSR is not running for each
pool operation, we move that check inside `with_pool_operation`.

Signed-off-by: lippirk <ben.anson@citrix.com>