CA-342935: Disengage guest balloon driver and reset PV features on soft reset
Keeping balloon target set confuses those kdump kernels with balloon memory
hotplug compiled in since the target value does not correspond to the memory
reserved and available for the crash kernel. That usually results in balloon
driver trying to hotplug the difference quickly going out of reservation.
Toolstack doesn't know how much memory is reserved for the crash kernel by
the guest and, therefore, is unable to set the target properly.

If the target is not set in xenstore then balloon driver will not be enabled
in kdump kernel and the issue is avoided. Additionally, reset the rest of
PV features that the first kernel already declared - those will be
re-declared by kdump kernel accordingly.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>