UPD-666 CA-344268: Fix timing issue in PBIS available check
If the lwsmd service has only just started when xapi checks to see
whether PBIS / AD is available, then removing the KRBTGT user from
the PBIS cache fails. This is not a problem, but we should try again
to remove KRBTGT from the cache rather than erroring out completely.

To remedy this, we move the call to `ad-cache --delete-user` inside
the retry loop. Additionally, the retry loop has been decoupled from the
checking

This was tested by running a thread which restarts lwsmd periodically,
and then restarting xapi:

```
$ restart_lwsmd () { while true; do systemctl restart lwsmd; sleep 3; done; }
$ restart_lwsmd &
$ systemctl restart xapi
```

Unfortunately this only catches the issue some of the time; after
restarting xapi potentially several times, you should see the following
in xensource.log:

Auth_signature.Auth_failure("The authentication request could not be handled")

Signed-off-by: lippirk <ben.anson@citrix.com>
Signed-off-by: Christian Lindig <christian.lindig@citrix.com>