XSI-795 CA-343951 fix Nvidia version parsing
The version of the host driver is read from the library file name and
used to decide if multi-vgpu is supported. The version used to 123.56
but now contains another dot and thus caused the parsing to fail. We now
split the string at the dots and compare them piece wise as numbers (not
as strings). This means 1 = 1.0 = 1.0.0 and so on and "" = 0.

The format of the string is not completely validated. We split the
string and pass the elements to int_of_string to extract the version
number.

Xapi's configuration marks a version as "123.4+" to indicate the all
versions 123.4 and later are supported. We find such a version but again
do not validate the format completely.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>