CA-341518: send error handshake during migration
Suppose we are migrating a VM from src -> dest. If dest's xenopsd fails
inbetween synchronisation point 3 and 4, and fails to remove the
temporary dest VM, then it wouldn't send the error handshake as
expected.

We fix this using `finally`, and also explicitly handle this error case
on the src xenopsd.

This improves the user's experience, because previously this class of
migration failures would hang until you restart the toolstack, whereas
now they fail promptly.

Signed-off-by: lippirk <ben.anson@citrix.com>