CP-34472 add ability to record login failures
We expose a function which tracks failures to either a hashtable or a
counter. The hastable is preferred, but the counter is used if we can't
discern any information about the client. This is designed to be wrapped
around API calls which 'login'.

You can get a report of all the failed logins (since last asking for the
report) by calling `get_failed_login_stats`.

Signed-off-by: lippirk <ben.anson@citrix.com>