Improve CDF_* flags handling for domain creation
The existing flags logic is a mess and extremely difficult to extend.  Rewrite
it from scratch.

Introduce the assert_capability helper which checks for the availability of
specific features.  Xen performs the same checks, but this will provide
something more useful than "internal error EINVAL" to the caller.

Provide comments to explain what the flags are, and how they interact.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
---
This cleanup is necessary because I need to add CDF_NESTED_VIRT handling for
Xen 4.15.  A couple of questions:

1) The hap clause isn't great, but I can't think of any nicer way of doing it.
   The assert_capability must not be called for PV VMs, because both
   CAP_HAP/Shadow will be missing if HVM is compiled out of Xen.  Whatever
   pattern we go with here is going to repeat for CDF_NESTED_VIRT.

2) HAP handling is confusing.  There is vm_info.hap which is handled/derrived
   separately from platformdata["hap"].  The result here is more consistent,
   but leaves vm_info.hap unused as far as I can tell.

3) make()'s vcpus and no_sharept looks suspiciously like they should be part
   of vm_info.