CP-35021 VM.suspend - assert support for NVidia cards
So far, support for VM.suspend for VMs with a NVidia vGPU is assumed
provided the host driver supports it in principle:

  let is_suspendable vgpu =
    Db.VGPU.get_type ~__context ~self:vgpu |> fun self ->
    Db.VGPU_type.get_implementation ~__context ~self |> function
    | `nvidia | `nvidia_sriov ->
        let pgpu = Db.VGPU.get_resident_on ~__context ~self:vgpu in
        Db.is_valid_ref __context pgpu
        && Db.PGPU.get_compatibility_metadata ~__context ~self:pgpu
           |> List.mem_assoc Xapi_gpumon.Nvidia.key
    | _ ->
        false

But this becomes more nuanced as it now depends on the mode of the vGPU.
In some modes suspend is not supported. There is no clear API support
for this in the NVidia host library (which is accessed via gpumon).
Nvidia suggested to test for support of suspend by checking that the VM
can be migrated to the local host according to their API check. This is
what this commit implements.

This assumes that we can start a VM with vGPU without running the same
test (as otherwise it would be pointless to run it again) and that it is
only executed for suspend and migrate,

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>