CP-35284 validate and install updated host certs at start
If a host certificate is present in the file system at xapi start:

* verify and install it if none was present
* check if the information about it is still valid
* verify, update, and install it if it has changed

Installing a new self-signed certificate at startup in a pool is
dangerous unless the host is the only member of the pool. Check the pool
size to verify this.

If verification fails when a valid host certificate is required, enter
emergency mode because communication will be impossible with an invalid
host certificate.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>