REQ-403 CP-34461 emergency disable tls verification
An escape hatch is needed in case communication breaks down in a pool.
Importantly verification will only be disabled for the host on which you
make this emergency call. This introduces a discrepancy between the pool
database and the local database, which can be fixed by executing
Pool.enable_tls_verification once again.

Signed-off-by: lippirk <ben.anson@citrix.com>