CA-347560: Call VM.import_metadata_async for MD updates
Backport of 5522490.

VM.import_metadata may now block in case there are current operations running
on the VM in xenopsd. For metadata updates, it is sufficient that they happen
eventually, but we do not need to wait for them to complete. Therefore, call
the new async version of the function and just ignore the resulting task.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>