CA-347560: Introduce VM_import_metadata atomic and queue the op when needed
So far, `VM.import_metadata` calls were always executed immediately as they
came in, introducing or overwriting a VM's metadata in xenopsd's internal DB,
even if another operation of the VM is currently being executed. This could
even mean that records of devices are _deleted_ while operations on those
devices are still in the queue to be executed. This leads to a `Does_not_exist`
error and eventually an ugly `INTERNAL_ERROR` reported to the API client.

XAPI calls `VM.import_metadata` liberally in order to synchronise the VM's
state in xenopsd wrt its own state (the events-from-xapi thread), and such
calls can occur at any point in time. This means that there is a constant risk
of xenopsd's operations being affected.

The solution implemented in this patch is to postpone a metadata import if the
VM is already know to xenopsd by queueing it as an atomic operation. This
ensures that it is executed after any other current operations.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>