CP-35280: Rework gencert validation for cert chains
Now when installing a certificate the trust anchors from stunnel are
used to validate the certificate. This means that additional checks are
done to the chains and intermediate certificates they contain.

Now hashes are enforced for the whole chain (sha2), the chains must be
in order from leaf to root certificate. Additionally, properties of
x509v3 like path len are enforced when checking the chains.

When validating a chain a list of root cert fingerprints that validate
the certificate are returned. This is needed to fill the database with
correct information.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>